Fix: Don't include original extensions in filenames for convert mode
track.info.name contains the whole "file stem" (i.e. name sans extension),
rather than just a partial name. So, it is safe to use it instead of just
taking the original input filename.

Todo: Still need to figure out how to do a copy instead of convert when the
original and target extensions are the same.